refactor(telegram): extract update classification
